# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Trita API
  version: "2026-08-23"
  description: |
    Trita platform belső API-leírója — kézzel karbantartott, a route-fájlok
    Zod-sémáiból származtatva. Auth: Clerk session cookie (böngésző-hívások),
    kivéve a jelölt publikus, token-alapú végpontokat.

    Csoportok (tags): Public · Assessment · Observer · Profile · Notifications
    · Feedback · Team · TeamRoles/Trust/PsychSafety · Org · Manager/Hiring
    · Admin · Inquiry · System.

    Generálás: a projekt sima Next.js route-handlereket használ (nincs
    dekorátor-réteg), ezért teljesen automatikus generátor nincs bekötve.
    Ajánlott irány: zod-openapi (a meglévő Zod-sémák annotálása) vagy
    next-rest-framework — lásd docs/security/security-audit-2026-07-22.md.
servers:
  - url: https://trita.io
  - url: http://localhost:3000
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk session cookie — a böngésző automatikusan küldi.
    cronBearer:
      type: http
      scheme: bearer
      description: CRON_SECRET Bearer tokenként (Vercel cron).
    svixSignature:
      type: apiKey
      in: header
      name: svix-signature
      description: Svix webhook aláírás (svix-id + svix-timestamp + svix-signature).
  schemas:
...
    Answer:
      type: object
      required: [questionId, value]
      properties:
        questionId: { type: integer, minimum: 1 }
        value: { type: integer, minimum: 1, maximum: 5 }
security:
  - clerkSession: []
...
paths:
...
  /api/assessment/submit:
    post:
      tags: [Assessment]
      summary: Önértékelő teszt beküldése (score-számítás)
      description: Rate-limitelt.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [testType, answers]
              properties:
                testType: { type: string, enum: [TRITAN] }
                answers:
                  type: array
                  items: { $ref: "#/components/schemas/Answer" }
      responses: { "200": { description: Eredmény-azonosító } }
  /api/assessment/claim-guest:
    post:
      tags: [Assessment]
      summary: Vendég-kitöltés (/try) átvétele belépés után
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  items: { $ref: "#/components/schemas/Answer" }
      responses: { "200": { description: OK } }
...
  /api/career/fit:
    post:
      tags: [Career]
      summary: Karrier-illeszkedés számítása (motor v2, klaszterezett eredmény)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                prefs:
                  type: object
                  description: "tengelyenként -1 | 0 | 1 (people, variety, autonomy, creation, pace, structure, setting)"
                leadIntent: { type: string, enum: [lead, expert, unsure] }
                eduLevel: { type: string, enum: [primary, secondary, vocational, higher], nullable: true }
                industries: { type: array, maxItems: 3, items: { type: string } }
                currentIndustry: { type: string, nullable: true }
                limit: { type: integer, minimum: 1, maximum: 40 }
                readyOnly: { type: boolean }
      responses:
        "200": { description: OK }
        "409": { description: NO_ASSESSMENT — nincs kitöltött önértékelés }
...
  /api/candidate/{token}/team-role:
    parameters:
      - { name: token, in: path, required: true, schema: { type: string } }
    post:
      tags: [Manager/Hiring]
      summary: "[PARKOLT: hiring] Jelölt csapatszerep-lépésének beküldése"
      security: []
      responses: { "200": { description: OK }, "404": { description: "FEATURE_PARKED" } }
...